fix(ads): use ::numeric cast on sum itself, not on coalesce result
coalesce(sum(numeric_col), 0) was failing because PG couldn't infer a
common type. Cast the sum to numeric first, then the literal 0 to
numeric, so the coalesce branches share the same type.

+++ server/src/routes/ads.ts
@@ -493,7 +493,7 @@ export function adsRoutes(db: Db): Router {
         platform: adsInsights.platform,
         impressions: sql<number>`coalesce(sum(${adsInsights.impressions}),0)::int`,
         clicks: sql<number>`coalesce(sum(${adsInsights.clicks}),0)::int`,
-        spend: sql<string>`coalesce(sum(${adsInsights.spend}),0)::numeric`,
+        spend: sql<string>`coalesce(sum(${adsInsights.spend})::numeric, 0::numeric)`,
         leads: sql<number>`coalesce(sum(${adsInsights.leads}),0)::int`,
         conversions: sql<number>`coalesce(sum(${adsInsights.conversions}),0)::int`,
         days: sql<number>`count(distinct ${adsInsights.date})::int`,
